[mediadb] parse picasa ini files
In order to maintain the linkage between the image files and what picasa
knows of them, we must also process them and make new fresh picasa ini
files if we make any changes in file locations or names.
Further, the metadata might be useful.

Secondary: added debug info on the date processing and debugged the date
regex logic.